Safari-609.1.11.1/LayoutTests/media/track/track-webvtt-tc007-cue-no-id-expected.txt

Tests empty cue identifiers (they are optional), including cues with "-->", which are treated as empty.



*** Testing text track 0
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].id == '') OK
EXPECTED (cues[0].startTime == '0') OK
EXPECTED (cues[0].endTime == '30.5') OK
EXPECTED (cues[0].text == 'Bear is Coming!!!!!') OK
EXPECTED (cues[1].id == '') OK
EXPECTED (cues[1].startTime == '31') OK
EXPECTED (cues[1].endTime == '60.5') OK
EXPECTED (cues[1].text == 'I said Bear is coming!!!!') OK
EXPECTED (cues[2].id == '') OK
EXPECTED (cues[2].startTime == '61') OK
EXPECTED (cues[2].endTime == '1200.5') OK
EXPECTED (cues[2].text == 'I said Bear is coming now!!!!') OK

*** Testing text track 1
EXPECTED (cues.length == '3') OK
EXPECTED (cues[0].id == '') OK
EXPECTED (cues[0].startTime == '0') OK
EXPECTED (cues[0].endTime == '30.5') OK
EXPECTED (cues[0].text == 'Bear is Coming!!!!!') OK
EXPECTED (cues[1].id == '') OK
EXPECTED (cues[1].startTime == '31') OK
EXPECTED (cues[1].endTime == '60.5') OK
EXPECTED (cues[1].text == 'I said Bear is coming!!!!') OK
EXPECTED (cues[2].id == '') OK
EXPECTED (cues[2].startTime == '61') OK
EXPECTED (cues[2].endTime == '1200.5') OK
EXPECTED (cues[2].text == 'I said Bear is coming now!!!!') OK
END OF TEST
